pyproject.toml
[tool.poetry]
name = "musicbucket-bot"
version = "0.1.0"
description = ""
authors = ["Pau Riera <pau.riera.forteza@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "==0.21.*"
python-telegram-bot = "==20.0"
requests = "==2.28.*"
aiohttp = "==3.8.*"
emoji = "==2.2.*"
sentry-sdk = "==1.12.1"

[tool.poetry.dev-dependencies]
pytest = "*"
ipdb = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
